+++ src/scripts/marker.js
@@ -1,11 +1,45 @@
-import { MarkerAnimation } from './markeranimation.js';
-import { Settings } from './settings.js';
-import { findTokenById, Flags, FlagScope, socketAction, socketName } from './utils.js';
+import {MarkerAnimation} from './markeranimation.js';
+import {Settings} from './settings.js';
+import {findTokenById, Flags, FlagScope, socketAction, socketName} from './utils.js';
 
 /**
  * Provides functionality for creating, moving, and animating the turn marker
  */
 export class Marker {
+
+    /**
+     * Deletes any tiles flagged as a 'Turn Marker' from the canvas
+     */
+    static async deleteTurnMarker() {
+        const to_delete = canvas.scene.getEmbeddedCollection('Tile')
+            .filter(tile => tile.flags.turnMarker)
+            .map(tile => tile._id);
+        if (!game.user.isGM) {
+            game.socket.emit(socketName, {
+                mode: socketAction.deleteTurnMarker,
+                tileData: to_delete.data
+            });
+        } else {
+            await canvas.scene.deleteEmbeddedEntity('Tile', to_delete);
+        }
+    }
+
+    /**
+     * Deletes any tiles flagged as a 'Deck Marker' from the canvas
+     */
+    static async deleteOnDeckMarker() {
+        const to_delete = canvas.scene.getEmbeddedCollection('Tile')
+            .filter(tile => tile.flags.deckMarker)
+            .map(tile => tile._id);
+        if (!game.user.isGM) {
+            game.socket.emit(socketName, {
+                mode: socketAction.deleteOnDeckMarker,
+                tileData: to_delete.data
+            });
+        } else {
+            await canvas.scene.deleteEmbeddedEntity('Tile', to_delete);
+        }
+    }
 
     /**
      * Places a new turn marker under the token specified, and if required, starts the animation
@@ -15,13 +49,12 @@ export class Marker {
      */
     static async placeTurnMarker(tokenId, markerId) {
         if (!markerId) {
-            this.clearAllMarkers();
+            await this.deleteTurnMarker();
 
-            if (Settings.getTurnMarkerEnabled()) {
+            if (Settings.getIsEnabled("turnmarker")) {
                 let token = findTokenById(tokenId);
-                let ratio = Settings.getRatio();
-                let dims = this.getImageDimensions(token);
-                let center = this.getImageLocation(token);
+                let dims = this.getImageDimensions(token, false, "turnmarker");
+                let center = this.getImageLocation(token, false, "turnmarker");
 
                 let newTile = new Tile({
                     img: Settings.getImagePath(),
@@ -33,7 +66,7 @@ export class Marker {
                     rotation: 0,
                     hidden: token.data.hidden,
                     locked: false,
-                    flags: { turnMarker: true }
+                    flags: {turnMarker: true}
                 });
 
                 let tile = await canvas.scene.createEmbeddedEntity('Tile', newTile.data);
@@ -43,19 +76,53 @@ export class Marker {
                 return null;
             }
         } else {
-            this.moveMarkerToToken(tokenId, markerId);
+            await this.moveMarkerToToken(tokenId, markerId, "turnmarker");
             return markerId;
         }
     }
+
+    static async placeOnDeckMarker(tokenId) {
+        await this.deleteOnDeckMarker();
+
+        if (Settings.getIsEnabled("deckmarker")) {
+            let token = findTokenById(tokenId);
+            let dims = this.getImageDimensions(token, false, "deckmarker");
+            let center = this.getImageLocation(token, false, "deckmarker");
+            let newTile = new Tile({
+                img: Settings.getOnDeckImagePath(),
+                width: dims.w,
+                height: dims.h,
+                x: center.x,
+                y: center.y,
+                z: 900,
+                rotation: 0,
+                hidden: token.data.hidden,
+                locked: false,
+                flags: {deckMarker: true}
+            });
+
+            if (game.user.isGM) {
+                await canvas.scene.createEmbeddedEntity('Tile', newTile.data);
+            }
+        }
+    }
+
 
     /**
      * Deletes any tiles flagged as a 'Start Marker' from the canvas
      */
     static async deleteStartMarker() {
-        for (let tile of canvas.scene.getEmbeddedCollection('Tile')) {
-            if (tile.flags.startMarker) {
-                await canvas.scene.deleteEmbeddedEntity('Tile', tile._id);
-            }
+        const to_delete = canvas.scene.getEmbeddedCollection('Tile')
+            .filter(tile => tile.flags.startMarker)
+            .map(tile => tile._id);
+        if (!game.user.isGM) {
+            game.socket.emit(socketName, {
+                mode: socketAction.deleteStartMarker,
+                tileData: to_delete.data
+            });
+        } else {
+            await canvas.scene.unsetFlag(FlagScope, Flags.startMarkerPlaced);
+            await canvas.scene.deleteEmbeddedEntity('Tile', to_delete);
         }
     }
 
@@ -64,7 +131,9 @@ export class Marker {
      * @param {String} tokenId - The ID of the token to place the start marker under
      */
     static async placeStartMarker(tokenId) {
-        if (Settings.getStartMarkerEnabled()) {
+        await this.deleteStartMarker();
+
+        if (Settings.getIsEnabled("startmarker")) {
             let token = findTokenById(tokenId);
             let dims = this.getImageDimensions(token);
             let center = this.getImageLocation(token);
@@ -78,17 +147,12 @@ export class Marker {
                 rotation: 0,
                 hidden: token.data.hidden,
                 locked: false,
-                flags: { startMarker: true }
+                flags: {startMarker: true}
             });
 
             if (game.user.isGM) {
-                canvas.scene.createEmbeddedEntity('Tile', newTile.data);
-                canvas.scene.setFlag(FlagScope, Flags.startMarkerPlaced, true);
-            } else {
-                game.socket.emit(socketName, {
-                    mode: socketAction.placeStartMarker,
-                    tileData: newTile.data
-                });
+                await canvas.scene.createEmbeddedEntity('Tile', newTile.data);
+                await canvas.scene.setFlag(FlagScope, Flags.startMarkerPlaced, true);
             }
         }
     }
@@ -98,11 +162,10 @@ export class Marker {
      * @param {String} tokenId - The ID of the token that the marker should be placed under
      * @param {String} markerId - The ID of the tile currently serving as the turn marker
      */
-    static async moveMarkerToToken(tokenId, markerId) {
+    static async moveMarkerToToken(tokenId, markerId, marker_type="turnmarker") {
         let token = findTokenById(tokenId);
-        let ratio = Settings.getRatio();
-        let dims = this.getImageDimensions(token);
-        let center = this.getImageLocation(token);
+        let dims = this.getImageDimensions(token, false, marker_type);
+        let center = this.getImageLocation(token, false, marker_type);
 
         await canvas.scene.updateEmbeddedEntity('Tile', {
             _id: markerId,
@@ -118,23 +181,19 @@ export class Marker {
      * Removes any existing turn marker and start marker tiles from the canvas
      */
     static async clearAllMarkers() {
-        let tiles = canvas.scene.getEmbeddedCollection('Tile');
-
-        for (var tile of tiles) {
-            if (tile.flags.turnMarker || tile.flags.startMarker) {
-                await canvas.scene.deleteEmbeddedEntity('Tile', tile._id);
-            }
-        }
+        await this.deleteTurnMarker();
+        await this.deleteStartMarker();
+        await this.deleteOnDeckMarker();
     }
 
     /**
      * Updates the tile image when the image path has changed
      */
-    static updateImagePath() {
+    static async updateImagePath() {
         if (game.user.isGM) {
             let tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
             if (tile) {
-                canvas.scene.updateEmbeddedEntity('Tile', {
+                await canvas.scene.updateEmbeddedEntity('Tile', {
                     _id: tile.id,
                     img: Settings.getImagePath()
                 });
@@ -143,62 +202,90 @@ export class Marker {
     }
 
     /**
-     * Completely resets the turn marker - deletes all tiles and stops any animation
-     * @param {Object} animator - The animator object
+     * Updates the tile image when the image path has changed
      */
-    static reset(animator) {
-        MarkerAnimation.stopAnimation(animator);
-        this.clearAllMarkers();
+    static async updateOnDeckImagePath() {
+        if (game.user.isGM) {
+            let tile = canvas.tiles.placeables.find(t => t.data.flags.deckMarker == true);
+            if (tile) {
+                await canvas.scene.updateEmbeddedEntity('Tile', {
+                    _id: tile.id,
+                    img: Settings.getOnDeckImagePath()
+                });
+            }
+        }
+    }
+
+    /**
+     * Completely resets the turn marker - deletes all tiles and stops any animation
+     */
+    static async reset() {
+        MarkerAnimation.stopAllAnimation();
+        await this.clearAllMarkers();
     }
 
     /**
      * Gets the proper dimensions of the marker tile taking into account the current grid layout
      * @param {object} token - The token that the tile should be placed under
      */
-    static getImageDimensions(token, ignoreRatio = false) {
-        let ratio = ignoreRatio ? 1 : Settings.getRatio();
+    static getImageDimensions(token, ignoreRatio = false, marker_type = "turnmarker") {
+        let ratio = ignoreRatio ? 1 : Settings.getRatio(marker_type);
         let newWidth = 0;
         let newHeight = 0;
 
         switch (canvas.grid.type) {
-            case 2: case 3: // Hex Rows
+            case 2:
+            case 3: // Hex Rows
                 newWidth = newHeight = token.h * ratio;
                 break;
-            case 4: case 5: // Hex Columns
+            case 4:
+            case 5: // Hex Columns
                 newWidth = newHeight = token.w * ratio;
                 break;
             default: // Gridless and Square
-                newWidth = token.w * ratio;
-                newHeight = token.h * ratio;
+                newWidth = this.getSmallerDimension(token.w, token.h) * ratio;
+                newHeight = this.getSmallerDimension(token.w, token.h) * ratio;
                 break;
         }
 
-        return { w: newWidth, h: newHeight };
+        return {w: newWidth, h: newHeight};
+    }
+
+    /**
+     * Returns the smaller dimension, so we can prevent lopsided markers when we have larger tokens on square/gridless
+     * @param width
+     * @param height
+     * @returns {*}
+     */
+    static getSmallerDimension(width, height) {
+        return width < height ? width : height;
     }
 
     /**
      * Gets the proper location of the marker tile taking into account the current grid layout
      * @param {object} token - The token that the tile should be placed under
      */
-    static getImageLocation(token, ignoreRatio = false) {
-        let ratio = ignoreRatio ? 1 : Settings.getRatio();
+    static getImageLocation(token, ignoreRatio = false, marker_type = "turnmarker") {
+        let ratio = ignoreRatio ? 1 : Settings.getRatio(marker_type);
         let newX = 0;
         let newY = 0;
 
         switch (canvas.grid.type) {
-            case 2: case 3: // Hex Rows
+            case 2:
+            case 3: // Hex Rows
                 newX = token.center.x - ((token.h * ratio) / 2);
                 newY = token.center.y - ((token.h * ratio) / 2);
                 break;
-            case 4: case 5: // Hex Columns
+            case 4:
+            case 5: // Hex Columns
                 newX = token.center.x - ((token.w * ratio) / 2);
                 newY = token.center.y - ((token.w * ratio) / 2);
                 break;
             default: // Gridless and Square
-                newX = token.center.x - ((token.w * ratio) / 2);
-                newY = token.center.y - ((token.h * ratio) / 2);
+                newX = token.center.x - ((this.getSmallerDimension(token.w, token.h) * ratio) / 2);
+                newY = token.center.y - ((this.getSmallerDimension(token.w, token.h) * ratio) / 2);
         }
 
-        return { x: newX, y: newY };
+        return {x: newX, y: newY};
     }
 }