+++ src/scripts/turnmarker.js
@@ -1,27 +1,18 @@
-import { Chatter } from './chatter.js';
-import { Marker } from './marker.js';
-import { MarkerAnimation } from './markeranimation.js';
-import { Settings } from './settings.js';
-import { renderUpdateWindow } from './updateWindow.js';
-import { firstGM, Flags, FlagScope, socketAction, socketName } from './utils.js';
+import {Chatter} from './chatter.js';
+import {Marker} from './marker.js';
+import {MarkerAnimation} from './markeranimation.js';
+import {Settings} from './settings.js';
+import {renderUpdateWindow} from './updateWindow.js';
+import {firstGM, Flags, FlagScope, socketAction, socketName, getNextTurn} from './utils.js';
 
-let animator;
-let markerId;
+
 let lastTurn = '';
 
-Hooks.on('ready', async () => {
+Hooks.once('init', () => {
     Settings.registerSettings();
-    let marker = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
-    if (marker && marker.id) {
-        markerId = marker.id;
-        let tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
-        tile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
-        tile.parent.sortChildren();
-        if (!game.paused && Settings.getShouldAnimate()) {
-            animator = MarkerAnimation.startAnimation(animator, markerId);
-        }
-    }
+});
 
+Hooks.once('ready', () => {
     if (game.user.isGM) {
         if (isNewerVersion(game.modules.get("turnmarker").data.version, Settings.getVersion())) {
             renderUpdateWindow();
@@ -30,31 +21,65 @@ Hooks.on('ready', async () => {
 
     game.socket.on(socketName, async (data) => {
         if (game.user.isGM) {
-            switch (data.mode) {
-                case socketAction.placeStartMarker:
-                    await canvas.scene.createEmbeddedEntity('Tile', data.tileData);
-                    canvas.scene.setFlag(FlagScope, Flags.startMarkerPlaced, true);
+            if (data) {
+                const to_delete = canvas.tiles.placeables.find(t => t.id === data[0]);
+                switch (data.mode) {
+                    case socketAction.deleteStartMarker:
+                        await canvas.scene.deleteEmbeddedEntity('Tile', to_delete);
+                        canvas.scene.setFlag(FlagScope, Flags.startMarkerPlaced, true);
+                        break;
+                    case socketAction.deleteTurnMarker:
+                        await canvas.scene.deleteEmbeddedEntity('Tile', to_delete);
+                        break;
+                    case socketAction.deleteOnDeckMarker:
+                        await canvas.scene.deleteEmbeddedEntity('Tile', to_delete);
+                        break;
+                }
+
             }
         }
     });
 });
 
-Hooks.on('createTile', (scene, tile) => {
-    if (tile.flags.turnMarker == true) {
-        markerId = tile._id;
-        tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
+Hooks.on('canvasReady', () => {
+    let deckTile = canvas.tiles.placeables.find(t => t.data.flags.deckMarker == true);
+    if (deckTile) {
+        deckTile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
+        deckTile.parent.sortChildren();
+        if (!game.paused && Settings.getShouldAnimate("deckmarker")) {
+            MarkerAnimation.startAnimation("deckmarker");
+        }
+    }
+
+    let tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
+    if (tile) {
         tile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
         tile.parent.sortChildren();
-        if (Settings.getShouldAnimate()) {
-            animator = MarkerAnimation.startAnimation(animator, markerId);
+        if (!game.paused && Settings.getShouldAnimate("turnmarker")) {
+            MarkerAnimation.startAnimation("turnmarker");
         }
     }
 });
 
-Hooks.on('preUpdateToken', async (scene, token) => {
-    if (game.combat != null && game.combat.turn != -1) {
-        if (token._id == game.combat.combatant.token._id && !canvas.scene.getFlag(FlagScope, Flags.startMarkerPlaced)) {
-            await Marker.placeStartMarker(game.combat.combatant.token._id);
+Hooks.on('createTile', (scene, data) => {
+    if (data.flags.turnMarker == true || data.flags.startMarker == true) {
+        const tile = canvas.tiles.placeables.find(t => t.id === data._id);
+        if (tile) {
+            if (data.flags.deckMarker == true) {
+                tile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
+                tile.parent.sortChildren();
+                if (!game.paused && Settings.getShouldAnimate("deckmarker")) {
+                    MarkerAnimation.startAnimation("deckmarker");
+                }
+            }
+            else if (data.flags.turnMarker == true) {
+                tile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
+                tile.parent.sortChildren();
+                if (!game.paused && Settings.getShouldAnimate("turnmarker")) {
+                    MarkerAnimation.startAnimation("turnmarker");
+                }
+            }
+            tile.renderable = isVisible(tile);
         }
     }
 });
@@ -63,24 +88,19 @@ Hooks.on('updateCombat', async (combat, update) => {
     // Clear out any leftovers, there seems to be a buggy instance where updateCombat is fired, when combat isn't
     // started nor, is a turn changed
     if (!combat.started) {
-        Marker.deleteStartMarker();
+        await Marker.deleteStartMarker();
     }
     if (combat.combatant) {
+        let nextTurn = getNextTurn(combat);
         if (update && lastTurn != combat.combatant._id && game.user.isGM && game.userId == firstGM()) {
             lastTurn = combat.combatant._id;
             if (combat && combat.combatant && combat.started) {
+                await Marker.placeStartMarker(combat.combatant.token._id);
+                createCombatDeckMarker(combat, nextTurn);
                 let tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
-                let result = await Marker.placeTurnMarker(combat.combatant.token._id, (tile && tile.id) || undefined);
-                if (result) {
-                    markerId = result.markerId;
-                    animator = result.animator;
-                }
-                if (Settings.getTurnMarkerEnabled()) {
-                    Marker.deleteStartMarker();
-                    canvas.scene.unsetFlag(FlagScope, Flags.startMarkerPlaced);
-                }
+                await Marker.placeTurnMarker(combat.combatant.token._id, (tile && tile.id) || undefined);
                 if (Settings.shouldAnnounceTurns() && !combat.combatant.hidden) {
-                    switch (Settings.getAnnounceActors()){
+                    switch (Settings.getAnnounceActors()) {
                         case 0:
                             Chatter.sendTurnMessage(combat.combatant);
                             break;
@@ -105,42 +125,131 @@ Hooks.on('updateCombat', async (combat, update) => {
 
 Hooks.on('deleteCombat', async () => {
     if (game.user.isGM) {
-        Marker.clearAllMarkers();
+        await Marker.clearAllMarkers();
     }
-    MarkerAnimation.stopAnimation(animator);
+    MarkerAnimation.stopAllAnimation();
 });
 
-Hooks.on('updateToken', (scene, updateToken, updateData) => {
+Hooks.on('updateToken', async (scene, updateToken, updateData) => {
+    /*
+    Moving preUpdateToken logic here, since pre hooks induce race conditions
+     */
+
+    // Do onDeck first, so current token will have higher Z-index
+    let deckTile = canvas.tiles.placeables.find(t => t.data.flags.deckMarker == true);
+    if (deckTile) {
+        if ((updateData.x || updateData.y || updateData.width || updateData.height || updateData.hidden) &&
+            (game && game.combat) &&
+            game.user.isGM && game.combat) {
+            let currentTurn = game.combat.turn
+            let nextTurn = currentTurn + 1;
+            if (nextTurn >= game.combat.turns.length) {
+                nextTurn = 0;
+            }
+            let nextToken = game.combat.turns[nextTurn].token
+            await Marker.moveMarkerToToken(nextToken._id, deckTile.id, "deckmarker");
+            deckTile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
+            deckTile.parent.sortChildren();
+        }
+    }
+
+
     let tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
     if (tile) {
         if ((updateData.x || updateData.y || updateData.width || updateData.height || updateData.hidden) &&
             (game && game.combat && game.combat.combatant && game.combat.combatant.tokenId == updateToken._id) &&
             game.user.isGM && game.combat) {
-            Marker.moveMarkerToToken(updateToken._id, tile.id);
+            await Marker.moveMarkerToToken(updateToken._id, tile.id, "turnmarker");
             tile.zIndex = Math.max(...canvas.tiles.placeables.map(o => o.zIndex)) + 1;
             tile.parent.sortChildren();
         }
     }
 });
 
-Hooks.on('updateTile', () => {
-    if (canvas.scene.data.tokenVision) {
-        let tile = canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true);
+function isVisible(tile) {
+    if (tile.data.hidden) {
+        return game.user.isGM;
+    }
+
+    if (!canvas.sight.tokenVision) {
+        return true;
+    }
+
+    if (tile._controlled) {
+        return true;
+    }
+
+    const combatant = canvas.tokens.placeables.find(t => t.id === game.combat.combatant.tokenId);
+
+    if (!combatant || combatant.data.hidden) {
+        return game.user.isGM;
+    }
+
+    if (combatant._controlled) {
+        return true;
+    }
+
+    let marker_type = "turnmarker";
+    if (tile.data.flags.startMarker) {
+        marker_type = "startmarker";
+    } else if (tile.data.flags.deckMarker) {
+        marker_type = "deckmarker";
+    }
+
+    const ratio = Settings.getRatio(marker_type);
+    const w = tile.data.width / ratio;
+    const h = tile.data.height / ratio;
+    const tolerance = Math.min(w, h) / 4;
+
+    return canvas.sight.testVisibility(tile.center, {tolerance, object: tile});
+}
+
+async function createCombatDeckMarker(combat, nextTurn) {
+    if (Settings.getDeckPlayersOnly()) {
+        if (combat.turns[nextTurn].actor.hasPlayerOwner) {
+            await Marker.placeOnDeckMarker(combat.turns[nextTurn].token._id).then(function () {
+                if (Settings.getShouldAnimate("deckmarker")) {
+                    MarkerAnimation.startAnimation("deckmarker");
+                }
+            });
+        } else {
+            await Marker.deleteOnDeckMarker();
+        }
+    } else {
+        await Marker.placeOnDeckMarker(combat.turns[nextTurn].token._id).then(function () {
+                if (Settings.getShouldAnimate("deckmarker")) {
+                    MarkerAnimation.startAnimation("deckmarker");
+                }
+        });
+    }
+}
+
+Hooks.on('updateTile', (entity, data, options, userId) => {
+    if (data.flags.turnMarker || data.flags.startMarker || data.flags.deckMarker) {
+        const tile = canvas.tiles.placeables.find(t => t.id === data._id);
         if (tile) {
-            let combatant = canvas.tokens.placeables.find(x => x.id == game.combat.combatant.tokenId);
-            if (combatant && !combatant.data.hidden) {
-                tile.visible = canvas.sight.testVisibility(combatant.center, { tolerance: canvas.dimensions.size / 4 });
-            }
+            tile.renderable = isVisible(tile);
         }
     }
 });
 
-Hooks.on('pauseGame', async (isPaused) => {
-    if (markerId && Settings.getShouldAnimate()) {
-        if (isPaused) {
-            MarkerAnimation.stopAnimation(animator);
-        } else {
-            animator = MarkerAnimation.startAnimation(animator, markerId);
+Hooks.on('sightRefresh', () => {
+    for (const tile of canvas.tiles.placeables) {
+        if (tile.data.flags.turnMarker || tile.data.flags.startMarker || tile.data.flags.deckMarker) {
+            tile.renderable = isVisible(tile);
         }
+    }
+});
+
+Hooks.on('pauseGame', (isPaused) => {
+    if (!isPaused) {
+        if (Settings.getShouldAnimate("turnmarker") && canvas.tiles.placeables.find(t => t.data.flags.turnMarker == true)) {
+            MarkerAnimation.startAnimation("turnmarker");
+        }
+        if (Settings.getShouldAnimate("deckmarker") && canvas.tiles.placeables.find(t => t.data.flags.deckMarker == true)) {
+            MarkerAnimation.startAnimation("deckmarker");
+        }
+    } else {
+        MarkerAnimation.stopAllAnimation();
     }
 });